move gethwrappers inside tools directory
